Modified all tests back to Junit4

+++ test/ResponseTreeMakerTest.java
@@ -5,18 +5,17 @@ import Entities.InterviewIntern;
 import java.io.FileNotFoundException;
 import java.util.ArrayList;
 import java.util.HashMap;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
+import org.junit.Test;
+import static org.junit.Assert.assertEquals;
 
 public class ResponseTreeMakerTest {
 
     @Test
     public void TestGenerateInternResponses() throws FileNotFoundException {
-        HashMap<String, Integer> skills = new HashMap<>();
-        skills.put("Confidence", 50);
-        skills.put("Teamwork", 50);
-        skills.put("Flexibility", 50);
+        HashMap<String, Double> skills = new HashMap<>();
+        skills.put("Confidence", 50.0);
+        skills.put("Teamwork", 50.0);
+        skills.put("Flexibility", 50.0);
         InterviewIntern intern = new InterviewIntern("Camille", 20, skills);
         ResponseTreeMaker rtm = new ResponseTreeMaker(intern);
         ResponseTree<ArrayList<String>> respTree = rtm.generateInternResponses();
@@ -25,10 +24,10 @@ public class ResponseTreeMakerTest {
 
     @Test
     public void TestAssignResponseToIntern() throws FileNotFoundException {
-        HashMap<String, Integer> skills = new HashMap<>();
-        skills.put("Confidence", 50);
-        skills.put("Teamwork", 50);
-        skills.put("Flexibility", 50);
+        HashMap<String, Double> skills = new HashMap<>();
+        skills.put("Confidence", 50.0);
+        skills.put("Teamwork", 50.0);
+        skills.put("Flexibility", 50.0);
         InterviewIntern intern = new InterviewIntern("Camille", 20, skills);
         ResponseTreeMaker rtm = new ResponseTreeMaker(intern);
         ResponseTree<ArrayList<String>> respTree = rtm.generateInternResponses();
